requirements.txt
flask==3.0.2
kafka-python==2.0.2
python-dotenv==1.0.1 